Moved state updates structs to separate file
