+++ radix-engine/src/track/mod.rs
@@ -1,5 +1,6 @@
 pub mod interface;
 pub mod legacy_state_updates;
+pub mod state_updates;
 pub mod track;
 pub mod utils;
 
@@ -8,4 +9,5 @@ mod test;
 
 pub use interface::*;
 pub use legacy_state_updates::*;
+pub use state_updates::*;
 pub use track::*;
